agentes_ia2/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "agente-sql"
version = "0.1.0"
description = "Agente IA para consultar SQL Server con lenguaje natural"
readme = "README.md"
requires-python = ">=3.11.9"
dependencies = [
    "openai",
    "python-dotenv",
    "pyodbc",
    "pandas",
    "sqlalchemy",
    "pydantic",
    "pydantic-settings",
    "tenacity",
    "fastapi",
    "uvicorn",
    "streamlit",
    "openpyxl",
    "reportlab",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "ruff"
]

[tool.setuptools.packages.find]
where = ["src"]